Bruk SystemOnlyPdlRestClient for å hente adresser i bolk
