+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/pdl/PdlRestClient.kt
@@ -4,12 +4,9 @@ import no.nav.familie.ba.sak.common.kallEksternTjeneste
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.DødsfallData
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.ForelderBarnRelasjon
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlBaseResponse
-import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlBostedsadresseOgDeltBostedPerson
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlDødsfallResponse
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlHentPersonResponse
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlOppholdResponse
-import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlPersonBolkRequest
-import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlPersonBolkRequestVariables
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlPersonRequest
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlPersonRequestVariables
 import no.nav.familie.ba.sak.integrasjoner.pdl.domene.PdlStatsborgerskapResponse
@@ -205,25 +202,6 @@ class PdlRestClient(
         ) {
             it.person!!.opphold
         }
-    }
-
-    fun hentBostedsadresseOgDeltBostedForPersoner(identer: List<String>): Map<String, PdlBostedsadresseOgDeltBostedPerson> {
-        val pdlPersonRequest =
-            PdlPersonBolkRequest(
-                variables = PdlPersonBolkRequestVariables(identer),
-                query = hentGraphqlQuery("bostedsadresse-og-delt-bosted"),
-            )
-
-        val pdlResponse: PdlBolkResponse<PdlBostedsadresseOgDeltBostedPerson> =
-            kallEksternTjeneste(
-                tjeneste = "pdl",
-                uri = pdlUri,
-                formål = "Hent bostedsadresse og delt bosted for personer",
-            ) {
-                postForEntity(pdlUri, pdlPersonRequest, httpHeaders())
-            }
-
-        return feilsjekkOgReturnerData(pdlResponse = pdlResponse)
     }
 
     fun hentUtenlandskBostedsadresse(aktør: Aktør): PdlUtenlandskAdresssePersonUtenlandskAdresse? {
